remove unused code from main; rename rtl to verilog
