+++ src/backend/mod.rs
@@ -1,5 +1,4 @@
 pub mod framework;
-// pub mod fsm;
 pub mod interpreter;
-pub mod rtl;
+pub mod verilog;
 pub mod traits;
